+++ game/Airside/Assets/Airside/Domain/AircraftType.cs
@@ -41,6 +41,7 @@ namespace Airside.Domain
         public static AircraftType Boeing7378 => AircraftCatalogue.Boeing7378.Type;
         public static AircraftType AirbusA321Neo => AircraftCatalogue.AirbusA321Neo.Type;
         public static AircraftType AirbusA350900 => AircraftCatalogue.AirbusA350900.Type;
+        public static AircraftType Boeing78710 => AircraftCatalogue.Boeing78710.Type;
 
         public bool CanReach(double legKm) => legKm <= PracticalRangeKm;